refactor exception throws with Optional

+++ src/main/java/ru/matveyelovskikh/naujavaspring/repository/impl/EventsDayRepository.java
@@ -8,6 +8,7 @@ import ru.matveyelovskikh.naujavaspring.repository.CrudRepository;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * Доступ к данным дневных событий, реализующий
@@ -17,7 +18,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 public class EventsDayRepository implements CrudRepository<EventsDayEntity, Long> {
 
     private final Map<Long, EventsDayEntity> eventsDayEntities;
-    private final AtomicInteger atomicInteger = new AtomicInteger();
+    private final AtomicLong atomicInteger = new AtomicLong();
 
     /**
      * Внедрение зависимости eventsDayEntities
@@ -30,7 +31,7 @@ public class EventsDayRepository implements CrudRepository<EventsDayEntity, Long
 
     @Override
     public void create(EventsDayEntity entity) {
-        entity.setId((long) atomicInteger.incrementAndGet());
+        entity.setId(atomicInteger.incrementAndGet());
         eventsDayEntities.put(entity.getId(), entity);
     }
 
@@ -46,7 +47,7 @@ public class EventsDayRepository implements CrudRepository<EventsDayEntity, Long
 
     @Override
     public void update(Long id, EventsDayEntity entity) {
-        entity.setId((long) atomicInteger.get());
+        entity.setId(atomicInteger.get());
         if (!id.equals(entity.getId())) {
             throw new IllegalArgumentException("ID не совпадают");
         }
